factor test utils out (#50)
* factor test utils out

* rm unused

* add rr store
