+++ Gopkg.lock
@@ -885,7 +885,6 @@
     "github.com/palantir/pkg/signals",
     "github.com/palantir/witchcraft-go-error",
     "github.com/palantir/witchcraft-go-logging/wlog",
-    "github.com/palantir/witchcraft-go-logging/wlog-zap",
     "github.com/palantir/witchcraft-go-logging/wlog/svclog/svc1log",
     "github.com/palantir/witchcraft-go-logging/wlog/wapp",
     "github.com/palantir/witchcraft-go-server/config",
@@ -903,6 +902,7 @@
     "k8s.io/apimachinery/pkg/api/resource",
     "k8s.io/apimachinery/pkg/apis/meta/v1",
     "k8s.io/apimachinery/pkg/labels",
+    "k8s.io/apimachinery/pkg/runtime",
     "k8s.io/apimachinery/pkg/selection",
     "k8s.io/apimachinery/pkg/types",
     "k8s.io/apimachinery/pkg/util/wait",
